@@ -1,5 +1,5 @@
 {
-  "name": "user",
+  "name": "your-angular-project",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
@@ -11,26 +11,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.0.3",
-    "@angular/cdk": "github:angular/cdk-builds",
-    "@angular/common": "~8.0.0",
-    "@angular/compiler": "~8.0.0",
-    "@angular/core": "~8.0.0",
-    "@angular/forms": "~8.0.0",
-    "@angular/material": "github:angular/material2-builds",
-    "@angular/platform-browser": "~8.0.0",
-    "@angular/platform-browser-dynamic": "~8.0.0",
-    "@angular/router": "~8.0.0",
-    "@auth0/angular-jwt": "^2.1.0",
+    "@angular/animations": "^8.1.0",
+    "@angular/cdk": "^8.0.2",
+    "@angular/common": "~8.1.0",
+    "@angular/compiler": "~8.1.0",
+    "@angular/core": "~8.1.0",
+    "@angular/forms": "~8.1.0",
+    "@angular/material": "^8.0.2",
+    "@angular/platform-browser": "~8.1.0",
+    "@angular/platform-browser-dynamic": "~8.1.0",
+    "@angular/router": "~8.1.0",
+    "@auth0/angular-jwt": "^2.1.1",
+    "@fortawesome/fontawesome-free": "^5.9.0",
+    "@types/chart.js": "^2.7.55",
+    "angular-bootstrap-md": "^7.5.4",
+    "chart.js": "^2.5.0",
+    "hammerjs": "^2.0.8",
     "rxjs": "~6.4.0",
     "tslib": "^1.9.0",
     "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.0",
-    "@angular/cli": "~8.0.1",
-    "@angular/compiler-cli": "~8.0.0",
-    "@angular/language-service": "~8.0.0",
+    "@angular-devkit/build-angular": "~0.801.0",
+    "@angular/cli": "~8.1.0",
+    "@angular/compiler-cli": "~8.1.0",
+    "@angular/language-service": "~8.1.0",
     "@types/node": "~8.9.4",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
