@@ -9,6 +9,6 @@ export class PostIdeaServiceService {
   constructor(private httpclient: HttpClient) { }
 
   postIdea(value): any{
-    return this.httpclient.post("http://localhost:8063/api/v1/addIdea",value).subscribe();
-      }
+    return this.httpclient.post("http://localhost:8087/api/v1/addIdea",value).subscribe();
+  }
 }
